+++ packages/fe-mockserver-core/src/mockdata/functionBasedMockData.ts
@@ -65,7 +65,7 @@ export type MockDataContributor = {
         actionData: any,
         keys: Record<string, any>,
         odataRequest: ODataRequest
-    ): Promise<object>;
+    ): Promise<object | undefined>;
     onAfterAction?(
         actionDefinition: Action,
         actionData: any,
@@ -274,7 +274,7 @@ export class FunctionBasedMockData extends FileBasedMockData {
         actionData: any,
         keys: Record<string, any>,
         odataRequest: ODataRequest
-    ): Promise<object> {
+    ): Promise<object | undefined> {
         if (this._mockDataFn?.executeAction) {
             return this._mockDataFn.executeAction(actionDefinition, actionData, keys, odataRequest);
         } else {