+++ packages/ui5-middleware-fe-mockserver/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/ui5-middleware-fe-mockserver
 
+## 2.1.65
+
+### Patch Changes
+
+-   @sap-ux/fe-mockserver-core@1.1.63
+
+## 2.1.64
+
+### Patch Changes
+
+-   Updated dependencies [0c1ec3e]
+    -   @sap-ux/fe-mockserver-core@1.1.62
+
+## 2.1.63
+
+### Patch Changes
+
+-   Updated dependencies [4050889]
+    -   @sap-ux/fe-mockserver-core@1.1.61
+
+## 2.1.62
+
+### Patch Changes
+
+-   Updated dependencies [4e3aca0]
+    -   @sap-ux/fe-mockserver-core@1.1.60
+
 ## 2.1.61
 
 ### Patch Changes