+++ packages/vocabularies-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/vocabularies-types
 
+## 0.7.2
+
+### Patch Changes
+
+-   7230a33: feat: allow to resolve entity type by their fqdn
+
+## 0.7.1
+
+### Patch Changes
+
+-   2df59c3: ActionImport and TypeDefinition added to type ServiceObjects
+
 ## 0.7.0
 
 ### Minor Changes